% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise_completude.R
\name{analise_completude}
\alias{analise_completude}
\title{Análise de Completude de Dados}
\usage{
analise_completude(dados)
}
\arguments{
\item{dados}{Um dataframe contendo as variáveis a serem analisadas.}
}
\value{
Retorna uma tabela com a completude de cada variável, incluindo um gráfico de barras horizontal referente aos Dados Ausentes.
}
\description{
Esta função calcula e avalia a completude de cada variável em um dataframe e gera um gráfico de barras horizontal. A completude é exibida junto com uma classificação baseada no escore de Romero & Cunha. Referência: ROMERO, Dalia E.; CUNHA, Cynthia Braga da. Avaliação da qualidade das variáveis sócio-econômicas e demográficas dos óbitos de crianças menores de um ano registrados no Sistema de Informações sobre Mortalidade do Brasil (1996/2001). Cadernos de Saúde Pública, v. 22, p. 673-681, 2006.
}
\examples{
# Supondo que você tenha um dataframe chamado dados_RHC_combinados,
# Use a função da seguinte forma:

analise_completude(dados_RHC_combinados)

OBSERVAÇÃO: Sem atribuir a nenhum objeto.
}
