% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcoes_RHCgen.R
\name{analise_completude_ano}
\alias{analise_completude_ano}
\title{Análise de Completude de Dados por Ano}
\usage{
analise_completude_ano(dados)
}
\arguments{
\item{dados}{Um dataframe contendo as variáveis a serem analisadas.}
}
\value{
Retorna uma tabela com a completude de cada variável por ano, incluindo uma legenda com a classificação de completude.
}
\description{
Esta função calcula e avalia a completude de cada variável por ano, distribuindo a proporção de completude pela variável Ano_do_Banco. A completude é exibida junto com uma classificação baseada no escore de Romero & Cunha. Referência: ROMERO, Dalia E.; CUNHA, Cynthia Braga da. Avaliação da qualidade das variáveis sócio-econômicas e demográficas dos óbitos de crianças menores de um ano registrados no Sistema de Informações sobre Mortalidade do Brasil (1996/2001). Cadernos de Saúde Pública, v. 22, p. 673-681, 2006."
}
\examples{
# Supondo que você tenha um dataframe chamado dados_RHC_combinados,
# Use a função da seguinte forma:

Nome_da_Sua_Tabela_de_Completude <- analise_completude_ano(dados_RHC_combinados)
}
