% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analise_qualidade_dos_dados.R
\name{analise_qualidade_dos_dados}
\alias{analise_qualidade_dos_dados}
\title{Análise de Qualidade dos Dados}
\usage{
analise_qualidade_dos_dados(data)
}
\arguments{
\item{data}{Um dataframe contendo os dados a serem analisados.}
}
\value{
Retorna uma mensagem indicando a conclusão da análise.
}
\description{
Esta função executa uma análise completa de qualidade dos dados, chamando várias funções que verificam completude, códigos inexistentes, idades inconsistentes, anos inconsistentes, datas inconsistentes e inconsistências entre câncer e sexo.
}
\examples{
# Exemplo de uso:
# analise_qualidade_dos_dados(dados_RHC_combinados)
}
