% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificacao_CID_0.R
\name{classificacao_CID_0}
\alias{classificacao_CID_0}
\title{Classifica as Neoplasias Malignas baseada na localização anatômica e tipo histológico conforme o CID-O}
\usage{
classificacao_CID_0(data)
}
\arguments{
\item{data}{Um dataframe contendo a coluna \code{Localizacao_Primaria_3D} com os códigos CID-0.}
}
\value{
Retorna um dataframe com uma nova coluna \code{Classificacao_CID_0} indicando a categoria da neoplasia maligna.
}
\description{
Esta função cria uma nova coluna em um dataframe com a classificação das neoplasias malignas baseada na localização anatômica e tipo histológico conforme o CID-O.
}
\details{
A Classificação Internacional de Doenças para Oncologia (CID-O) e a Classificação Internacional de Doenças (CID-10) agrupam os cânceres com base na localização anatômica e no tipo histológico.
}
\examples{
# Exemplo de uso:
# dados_RHC_combinados <- classificacao_CID_0(dados_RHC_combinados)
}
