% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcoes_Pacote_RHCgen.R
\name{construir_banco}
\alias{construir_banco}
\title{Construir Banco de Dados Automaticamente}
\usage{
construir_banco(data)
}
\arguments{
\item{none}{Não são necessários parâmetros externos, pois a função opera com base no conteúdo do diretório especificado.}
}
\value{
Retorna o dataframe processado com todos os ajustes iniciais para posterior análise.
}
\description{
Esta função aplica uma série de ajustes a um dataframe para construir um banco de dados finalizado. As etapas incluem ler e juntar os arquivo DBF, renomear colunas, modificar tipos de variáveis, recodificar variáveis, renomear siglas dos estados, renomear códigos CID de 3 e 4 dígitos, renomear códigos CNES e Tipo Histolófico e adicionar novas variáveis ao final do dataframe.
}
\examples{
# Supondo que você tenha no diretório de trabalho arquivos DBF do RHC que permanecem com o mesmo nome que vieram do Integrador,
# Use a função da seguinte forma:

dados_RHC_combinados <- construir_banco()
# Se nenhum arquivo for encontrado ou se houver erros ao juntar os arquivos, dados_combinados será NULL.
# Caso contrário, você terá um dataframe com todos os dados ajustados de forma automática.
}
