% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcoes_Pacote_RHCgen.R
\name{filtrar_banco}
\alias{filtrar_banco}
\title{Filtrar Dados com Base em Vários Critérios}
\usage{
filtrar_banco(
  dados = Seu_data_frame_aqui,
  cid3digitos = NULL,
  cid4digitos = NULL,
  ano_inicio_triagem = NULL,
  ano_fim_triagem = NULL,
  ano_inicio_diagnostico = NULL,
  ano_fim_diagnostico = NULL,
  ano_inicio_primeira_consulta = NULL,
  ano_fim_primeira_consulta = NULL,
  idade_inicio = NULL,
  idade_fim = NULL,
  tipo_de_caso = NULL,
  sexo = NULL,
  estado_residencia = NULL,
  uf_unidade_hospital = NULL,
  primeiro_tratamento_hospital = NULL,
  ano_inicio_tratamento = NULL,
  ano_fim_tratamento = NULL,
  origem_do_encaminhamento = NULL
)
}
\arguments{
\item{dados}{Um dataframe que contém as variáveis a serem filtradas.}

\item{cid3digitos}{Um vetor de códigos CID-3 para filtrar a coluna \code{Localizacao_Primaria_3D}. Padrão é NULL.}

\item{cid4digitos}{Um vetor de códigos CID-4 para filtrar a coluna \code{Localizacao_Primaria_4D}. Padrão é NULL.}

\item{ano_inicio_triagem}{Ano inicial para filtrar a coluna \code{Ano_Triagem}. Padrão é NULL.}

\item{ano_fim_triagem}{Ano final para filtrar a coluna \code{Ano_Triagem}. Padrão é NULL.}

\item{ano_inicio_diagnostico}{Ano inicial para filtrar a coluna \code{Ano_Primeiro_Diagnostico}. Padrão é NULL.}

\item{ano_fim_diagnostico}{Ano final para filtrar a coluna \code{Ano_Primeiro_Diagnostico}. Padrão é NULL.}

\item{ano_inicio_primeira_consulta}{Ano inicial para filtrar a coluna \code{Ano_Primeira_Consulta}. Padrão é NULL.}

\item{ano_fim_primeira_consulta}{Ano final para filtrar a coluna \code{Ano_Primeira_Consulta}. Padrão é NULL.}

\item{idade_inicio}{Idade inicial para filtrar a coluna \code{Idade}. Padrão é NULL.}

\item{idade_fim}{Idade final para filtrar a coluna \code{Idade}. Padrão é NULL.}

\item{tipo_de_caso}{Filtro para a coluna \code{Tipo_de_Caso} (valores possíveis: "Analítico", "Não Analítico"). Padrão é NULL.}

\item{sexo}{Filtro para a coluna \code{Sexo} (valores possíveis: "Masculino", "Feminino"). Padrão é NULL.}

\item{estado_residencia}{Filtro para a coluna \code{Estado_Residencia} com siglas dos estados. Padrão é NULL.}

\item{uf_unidade_hospital}{Filtro para a coluna \code{UF_Unidade_Hospital} com siglas dos estados. Padrão é NULL.}

\item{primeiro_tratamento_hospital}{Filtro para a coluna \code{Primeiro_Tratamento_Hospital}. Padrão é NULL.}

\item{ano_inicio_tratamento}{Ano inicial para filtrar a coluna \code{Ano_Inicio_Tratamento}. Padrão é NULL.}

\item{ano_fim_tratamento}{Ano final para filtrar a coluna \code{Ano_Inicio_Tratamento}. Padrão é NULL.}

\item{origem_do_encaminhamento}{Filtro para a coluna \code{Origem_do_Encaminhamento}. Padrão é NULL.}
}
\value{
Retorna o dataframe filtrado com base nos critérios fornecidos.
}
\description{
Esta função filtra um dataframe com base em critérios específicos fornecidos pelo usuário, incluindo códigos CID-3, códigos CID-4, anos de triagem, anos de diagnóstico, anos da primeira consulta, idade, tipo de caso, sexo, estado de residência, unidade hospitalar, primeiro tratamento hospitalar, ano de início do tratamento e origem do encaminhamento.
}
\examples{
# Filtrar o banco de dados
DADOS_FILTRADOS <- filtrar_banco(dados = dados_RHC_combinados,
                                 cid3digitos = "C44",  # Filtra os dados pela coluna `Localizacao_Primaria_3D` usando o código CID-3 "C44"
                                 cid4digitos = "C44.0",  # Filtra os dados pela coluna `Localizacao_Primaria_4D` usando o código CID-4 "C44.0"
                                 ano_inicio_triagem = 2018,  # Filtra os dados pela coluna `Ano_Triagem` entre os anos de 2018 e 2020
                                 ano_fim_triagem = 2020,
                                 ano_inicio_diagnostico = 2015,  # Filtra os dados pela coluna `Ano_Primeiro_Diagnostico` entre os anos de 2015 e 2019
                                 ano_fim_diagnostico = 2019,
                                 ano_inicio_primeira_consulta = 2016,  # Filtra os dados pela coluna `Ano_Primeira_Consulta` entre os anos de 2016 e 2020
                                 ano_fim_primeira_consulta = 2020,
                                 idade_inicio = 20,  # Filtra os dados pela coluna `Idade` entre 20 e 80 anos
                                 idade_fim = 80,
                                 tipo_de_caso = "Analítico",  # Filtra os dados pela coluna `Tipo_de_Caso`, selecionando "Analítico"
                                 sexo = NULL,  # Não filtra pelo sexo (NULL)
                                 estado_residencia = c("RJ", "MG"),  # Filtra os dados pela coluna `Estado_Residencia`, selecionando "RJ" e "MG"
                                 uf_unidade_hospital = c("RJ", "MG"),  # Filtra os dados pela coluna `UF_Unidade_Hospital`, selecionando "RJ" e "MG"
                                 primeiro_tratamento_hospital = "Cirurgia",  # Filtra os dados pela coluna `Primeiro_Tratamento_Hospital`, selecionando "Cirurgia"
                                 ano_inicio_tratamento = 2014,  # Filtra os dados pela coluna `Ano_Inicio_Tratamento` entre os anos de 2014 e 2019
                                 ano_fim_tratamento = 2019,
                                 origem_do_encaminhamento = "SUS")  # Filtra os dados pela coluna `Origem_do_Encaminhamento`, selecionando "SUS"
}
