% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcoes_Pacote_RHCgen.R
\name{Funcoes_Pacote}
\alias{Funcoes_Pacote}
\title{Funções para Manipulação de Dados do RHC}
\description{
Funções para carregar e combinar arquivos RHC, construir bases de dados, filtrar dados, mapear siglas de estados, modificar tipos de variáveis, recodificar variáveis e renomear colunas.
}
\details{
Este pacote foi desenvolvido para facilitar a manipulação e análise de dados do RHC, fornecendo um conjunto abrangente de funções para diversas tarefas relacionadas à preparação e limpeza de dados.
}
\author{
\strong{Autor:} Anderson Lineu \email{andersonlineu@gmail.com}
}
