% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcoes_Pacote_RHCgen.R
\name{modificar_tipo_variavel}
\alias{modificar_tipo_variavel}
\title{Modificar Tipos de Variáveis em um DataFrame}
\usage{
modificar_tipo_variavel(data)
}
\arguments{
\item{data}{Um dataframe que contém as variáveis a serem convertidas.}
}
\value{
Retorna um dataframe com as variáveis convertidas para os tipos apropriados conforme definido.
}
\description{
Esta função converte todas as variáveis de um dataframe para o tipo character, exceto as colunas de data, que são convertidas primeiro para character e depois para data.
A coluna \code{Idade} é convertida para o tipo numérico.
}
\examples{
# Escreva o nome do dataframe e execute a função. Se seu dataframe for "dados_RHC_combinados", use a funçao como: 

dados_RHC_combinados <- modificar_tipo_variavel(dados_RHC_combinados)
}
