% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ordeando.R
\name{ordenando_colunas}
\alias{ordenando_colunas}
\title{Ajusta Colunas em uma Ordem Específica}
\usage{
ordenando_colunas(dados)
}
\arguments{
\item{dados}{Um dataframe contendo as colunas a serem ajustadas.}
}
\value{
Retorna um dataframe com as colunas ajustadas na ordem especificada.
}
\description{
Esta função ajusta as colunas do dataframe fornecido em uma ordem específica.
}
\examples{
# Escreva o nome do dataframe e execute a função. Se seu dataframe for "dados_RHC_combinados", use a função como:

dados_RHC_combinados <- ordenando_colunas(dados_RHC_combinados)
}
