% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcoes_Pacote_RHCgen.R
\name{recodificar_variaveis}
\alias{recodificar_variaveis}
\title{Recodificar Variáveis em um DataFrame}
\usage{
recodificar_variaveis(data)
}
\arguments{
\item{data}{Um dataframe que contém as variáveis a serem recodificadas.}
}
\value{
Retorna um dataframe com as variáveis recodificadas.
}
\description{
Esta função recodifica variáveis categóricas em um dataframe com base no dicionário de variáveis predefinido pelo RHC. As variáveis são transformadas em fatores com níveis específicos para melhorar a legibilidade e consistência dos dados.
}
\examples{
# Escreva o nome do dataframe e execute a função. Se seu dataframe for "dados_RHC_combinados", use a funçao como: 

dados_RHC_combinados <- recodificar_variaveis(dados_RHC_combinados)
}
