% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcoes_Pacote_RHCgen.R
\name{renomear_CID_3digitos}
\alias{renomear_CID_3digitos}
\title{Renomear CID de 3 Dígitos}
\usage{
renomear_CID_3digitos(dados)
}
\arguments{
\item{dados}{Um dataframe contendo a coluna \code{Localizacao_Primaria_3D}.}
}
\value{
Retorna um dataframe com os nomes completos dos CID  de 3 dígitos adicionados em uma nova coluna.
}
\description{
Esta função mapeia os códigos CID de 3 dígitos para seus nomes completos na coluna \code{Localizacao_Primaria_3D} de um dataframe. Os nomes completos dos CID  de 3 dígitos são adicionados em uma nova coluna \code{CID3d}.
}
\examples{
# Escreva o nome do dataframe e execute a função. Se seu dataframe for "dados_RHC_combinados", use a funçao como: 

dados_RHC_combinados <- renomear_CID_3digitos(dados_RHC_combinados)
}
