% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renomear_clinica.R
\name{renomear_clinica}
\alias{renomear_clinica}
\title{Converte os Códigos das Clínicas de Atendimento e Tratamento de Um Dataframe}
\usage{
renomear_clinica(dados)
}
\arguments{
\item{dados}{Um dataframe contendo as colunas \code{Clinica_Atendimento} e \code{Clinica_Tratamento}.}
}
\value{
Retorna um dataframe com os nomes completos das clínicas de atendimento e tratamento adicionados em novas colunas.
}
\description{
Esta função mapeia os códigos das clínicas para seus nomes completos nas colunas \code{Clinica_Atendimento} e \code{Clinica_Tratamento} de um dataframe. Os nomes completos das clínicas são adicionados em novas colunas \code{Nome_Clinica_Atendimento} e \code{Nome_Clinica_Tratamento}.
}
\examples{
# Escreva o nome do dataframe e execute a função. Se seu dataframe for "dados_RHC_combinados", use a função como:

dados_RHC_combinados <- renomear_clinica(dados_RHC_combinados)
}
