% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcoes_RHCgen.R
\name{renomear_CNES}
\alias{renomear_CNES}
\title{Converte os Códigos CNES para os Nomes Completos.}
\usage{
renomear_CNES(dados)
}
\arguments{
\item{dados}{Um dataframe contendo a coluna \code{CNES_Hospital}.}
}
\value{
Retorna um dataframe com os nomes completos dos estabelecimentos adicionados em uma nova coluna.
}
\description{
Esta função mapeia os códigos CNES para os nomes completos dos estabelecimentos na coluna \code{CNES_Hospital} de um dataframe. Os nomes completos dos estabelecimentos são adicionados em uma nova coluna \code{Estabelecimento_Hospitalar}.
}
\examples{
# Escreva o nome do dataframe e execute a função. Se seu dataframe for "dados_RHC_combinados", use a função como:

dados_RHC_combinados <- renomear_CNES(dados_RHC_combinados)
}
