% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Municipios.R
\name{renomear_codigo_municipio_hospital}
\alias{renomear_codigo_municipio_hospital}
\title{Converte os Códigos dos Municípios para Nomes Completos}
\usage{
renomear_codigo_municipio_hospital(dados)
}
\arguments{
\item{dados}{Um dataframe contendo a coluna \code{Municipio_Unidade_Hospital}.}
}
\value{
Retorna um dataframe com os nomes completos dos municípios adicionados em uma nova coluna.
}
\description{
Esta função mapeia os códigos dos municípios brasileiros para seus nomes completos na coluna \code{Municipio_Unidade_Hospital} de um dataframe. Os nomes completos dos municípios são adicionados em uma nova coluna \code{Nome_Municipio_Hospital}.
}
\examples{
# Escreva o nome do dataframe e execute a função. Se seu dataframe for "dados_RHC_combinados", use a função como:

dados_RHC_combinados <- renomear_codigo_municipio_hospital(dados_RHC_combinados)
}
