% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renomear_colunas.R
\name{renomear_colunas}
\alias{renomear_colunas}
\title{Renomeia as Colunas de um DataFrame}
\usage{
renomear_colunas(data)
}
\arguments{
\item{data}{Um dataframe que contém as colunas a serem renomeadas.}
}
\value{
Um dataframe com as colunas renomeadas conforme especificado na lista de mapeamentos.
}
\description{
Esta função renomeia as colunas de um dataframe utilizando uma lista predefinida de mapeamentos entre os nomes originais das colunas e seus novos nomes descritivos. O objetivo é melhorar a legibilidade. A função verifica cada coluna listada no mapeamento e, se encontrada no dataframe, substitui seu nome pelo novo nome especificado. As colunas que não estão presentes no dataframe são ignoradas, e uma mensagem é exibida para cada coluna processada, indicando se a renomeação foi bem-sucedida ou não.
}
