% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renomear_estadiamento_clinico.R
\name{renomear_estadiamento_clinico}
\alias{renomear_estadiamento_clinico}
\title{Converte os Códigos do Estadiamento Clínico de Um Dataframe}
\usage{
renomear_estadiamento_clinico(dados)
}
\arguments{
\item{dados}{Um dataframe contendo a coluna \code{Estadiamento_Clinico}.}
}
\value{
Retorna um dataframe com os nomes do Estadiamento Clínico gerais adicionados em uma nova coluna.
}
\description{
Esta função renomeia o Estadiamento Clínico para a Classificação mais geral. Os nomes do Estadiamento Clínico são adicionados em uma nova coluna \code{Nome_Estadiamento_Clinico}.
}
\examples{
# Escreva o nome do dataframe e execute a função. Se seu dataframe for "dados_RHC_combinados", use a função como:

dados_RHC_combinados <- renomear_estadiamento_clinico(dados_RHC_combinados)
}
