% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcoes_Pacote_RHCgen.R
\name{renomear_siglas_estados}
\alias{renomear_siglas_estados}
\title{Renomear Siglas de Estados para Nomes Completos}
\usage{
renomear_siglas_estados(dados)
}
\arguments{
\item{dados}{Um dataframe contendo as colunas \code{Estado_Residencia} e \code{UF_Unidade_Hospital}.}
}
\value{
Retorna um dataframe com os nomes completos dos estados adicionados em novas colunas.
}
\description{
Esta função utiliza as siglas dos estados brasileiros das variáveis \code{Estado_Residencia} e \code{UF_Unidade_Hospital} de um dataframe e cria novas colunas com os seus nomes completos. Os nomes completos dos estados são adicionados em novas colunas \code{Nome_Estado_Residencia} e \code{Nome_Estado_Hospital}.
}
\examples{
# Escreva o nome do dataframe e execute a função. Se seu dataframe for "dados_RHC_combinados", use a funçao como: 

dados_RHC_combinados <- renomear_siglas_estados(dados_RHC_combinados)
}
