% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcoes_RHCgen.R
\name{renomear_tipo_histologico}
\alias{renomear_tipo_histologico}
\title{Renomear Tipo Histológico}
\usage{
renomear_tipo_histologico(dados)
}
\arguments{
\item{dados}{Um dataframe contendo a coluna \code{Tipo_Histologico}.}
}
\value{
Retorna um dataframe com os nomes completos dos tipos histológicos adicionados em uma nova coluna.
}
\description{
Esta função mapeia os códigos de Tipo Histológico para os nomes completos na coluna \code{Tipo_Histologico} de um dataframe. Os nomes completos dos tipos histológicos são adicionados em uma nova coluna \code{Tipo_Histologico_Completo}.
}
\examples{
# Escreva o nome do dataframe e execute a função. Se seu dataframe for "dados_RHC_combinados", use a função como:

dados_RHC_combinados <- renomear_tipo_histologico(dados_RHC_combinados)
}
