% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHCgen-package.R
\docType{package}
\name{RHCgen-package}
\alias{RHCgen-package}
\alias{RHCgen}
\title{RHCgen: Um Pacote R para Automação, Padronização e Avaliação da Qualidade dos Dados do Registro Hospitalar de Câncer no Brasil}
\description{
O RHCgen é um pacote desenvolvido em R para automatizar e padronizar o processamento de dados do Registro Hospitalar de Câncer (RHC) no Brasil. O pacote implementa um pipeline completo de extração, transformação e estruturação dos dados, incluindo integração de arquivos, harmonização estrutural, padronização semântica e avaliação da qualidade dos registros. Sua concepção responde à heterogeneidade histórica dos dados do RHC, permitindo a geração reprodutível de bases analíticas a partir de dados brutos.
}
\details{
Veja o guia passo a passo sobre como baixar os arquivos DBF no IntegradorRHC: \code{\link{Baixar_Arquivos_DBF}}
}
\seealso{
Tutorial: \code{\link{Baixar_Arquivos_DBF}}

Reportar Bugs: \url{https://github.com/andersonlineu/RHCgen/issues}
}
\author{
Anderson Lineu

Email: \email{andersonlineu@gmail.com}

Lattes: \url{http://lattes.cnpq.br/9727490230028203}

Orcid iD: \url{https://orcid.org/0000-0002-1703-9310}

Links úteis:

IntegradorRHC:  \url{https://irhc.inca.gov.br/RHCNet/}

Dicionário de Variáveis: \url{https://irhc.inca.gov.br/RHCNet/documentosTabWin.action?doc=dicionario.dados}

Também veja o Dicionário de Variáveis com o nome e a descrição após a criação de todo o banco de dados: \code{\link{Dicionario_Dados}}

Escore de Completude:  \url{https://www.scielo.br/j/csp/a/HXqrdksBsMrr4R9Ydnnmqcf/?format=pdf&lang=pt}
}
