% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verificar_anos_inconsistentes.R
\name{verificar_anos_inconsistentes}
\alias{verificar_anos_inconsistentes}
\title{Verifica Inconsistências Entre Anos em um Dataframe}
\usage{
verificar_anos_inconsistentes(data)
}
\arguments{
\item{data}{Um dataframe contendo as colunas "Ano_Primeiro_Diagnostico", "Ano_Primeira_Consulta", "Ano_Triagem" e "Ano_Inicio_Tratamento".}
}
\value{
Retorna um dataframe com o quantitativo e a proporção de inconsistências por verificação.
}
\description{
Esta função verifica se há inconsistências na ordem cronológica dos anos fornecidos em um dataframe. As verificações incluem:
\itemize{
\item "Ano_Primeiro_Diagnostico" deve ser igual ou posterior a "Ano_Primeira_Consulta"
\item "Ano_Primeiro_Diagnostico" deve ser igual ou posterior a "Ano_Triagem"
\item "Ano_Inicio_Tratamento" deve ser igual ou posterior a "Ano_Primeira_Consulta"
\item "Ano_Inicio_Tratamento" deve ser igual ou posterior a "Ano_Triagem"
\item "Ano_Inicio_Tratamento" deve ser igual ou posterior a "Ano_Primeiro_Diagnostico"
A função retorna um dataframe com o quantitativo e a proporção de inconsistências por verificação.
}
}
\examples{
# Escreva o nome do dataframe e execute a função. Se seu dataframe for "dados_RHC_combinados", use a função como:

resultados_anos_inconsistentes <- verificar_anos_inconsistentes(dados_RHC_combinados)
}
