% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verificar_cancer_sexo_inconsistentes.R
\name{verificar_cancer_sexo_inconsistentes}
\alias{verificar_cancer_sexo_inconsistentes}
\title{Verifica Inconsistências Entre CID-O e Sexo em um Dataframe}
\usage{
verificar_cancer_sexo_inconsistentes(data)
}
\arguments{
\item{data}{Um dataframe contendo as colunas "Localizacao_Primaria_3D" e "Sexo".}
}
\value{
Retorna um dataframe com o quantitativo e a proporção de inconsistências por verificação.
}
\description{
Esta função verifica se os CID-O dos cânceres relacionados ao sexo estão de acordo com a variável "Sexo". As verificações incluem:
\itemize{
\item CID-O de cânceres femininos (C53, C54, C55, C56, C57, C51, C52) devem corresponder ao sexo "Feminino"
\item CID-O de cânceres masculinos (C60, C61, C62) devem corresponder ao sexo "Masculino"
A função retorna um dataframe com o quantitativo e a proporção de inconsistências por verificação.
}
}
\examples{
# Escreva o nome do dataframe e execute a função. Se seu dataframe for "dados_RHC_combinados", use a função como:

resultados_cancer_sexo_inconsistentes <- verificar_cancer_sexo_inconsistentes(dados_RHC_combinados)
}
