% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verificar_codigos_inexistentes.R
\name{verificar_codigos_inexistentes}
\alias{verificar_codigos_inexistentes}
\title{Verifica Casos com Códigos Inexistentes em um Dataframe}
\usage{
verificar_codigos_inexistentes(data)
}
\arguments{
\item{data}{Um dataframe contendo as variáveis a serem verificadas.}
}
\value{
Retorna um dataframe com o quantitativo e a proporção de códigos inexistentes por variável.
}
\description{
Esta função verifica se há códigos inexistentes nas variáveis de um dataframe. As variáveis e seus códigos esperados são definidos dentro da função. A função retorna um dataframe com o quantitativo e a proporção de códigos inexistentes por variável.
}
\examples{
# Escreva o nome do dataframe e execute a função. Se seu dataframe for "dados_RHC_combinados", use a função como:

resultados_codigos_inexistentes <- verificar_codigos_inexistentes(dados_RHC_combinados)
}
