% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verificar_idades_inconsistentes.R
\name{verificar_idades_inconsistentes}
\alias{verificar_idades_inconsistentes}
\title{Verifica Idades Inconsistentes em um Dataframe}
\usage{
verificar_idades_inconsistentes(data)
}
\arguments{
\item{data}{Um dataframe contendo a coluna \code{Idade}.}
}
\value{
Retorna um dataframe com o quantitativo e a proporção de idades inválidas.
}
\description{
Esta função verifica se há idades inválidas (menores que zero ou maiores que 150 anos) na coluna \code{Idade} de um dataframe. A função retorna um dataframe com o quantitativo e a proporção de idades inválidas.
}
\examples{
# Escreva o nome do dataframe e execute a função. Se seu dataframe for "dados_RHC_combinados", use a função como:

resultados_idades_invalidas <- verificar_idades_invalidas(dados_RHC_combinados)
}
