% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcoes_Pacote_RHCgen.R
\name{renomear_CID_4digitos}
\alias{renomear_CID_4digitos}
\title{Renomear CID  de 4 Dígitos em um DataFrame}
\usage{
renomear_CID_4digitos(dados)
}
\arguments{
\item{dados}{Um dataframe contendo a coluna \code{Localizacao_Primaria_4D}.}
}
\value{
Retorna um dataframe com os nomes completos dos CID  de 4 dígitos adicionados em uma nova coluna.
}
\description{
Esta função mapeia os códigos CID  de 4 dígitos para seus nomes completos na coluna \code{Localizacao_Primaria_4D} de um dataframe. Os nomes completos dos CID de 4 dígitos são adicionados em uma nova coluna \code{CID4d}.
}
\examples{
# Escreva o nome do dataframe e execute a função. Se seu dataframe for "dados_RHC_combinados", use a funçao como: 

dados_RHC_combinados <- renomear_CID_4digitos(dados_RHC_combinados)
}
