% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHCgen-package.R
\docType{package}
\name{RHCgen-package}
\alias{RHCgen-package}
\alias{RHCgen}
\title{RHCgen: Pacote para construção de Banco de Dados de registros de câncer do RHC}
\description{
O RHCgen é um pacote desenvolvido para a linguagem de programação R, concebido com o objetivo de automatizar e padronizar o processamento de dados provenientes do Registro Hospitalar de Câncer (RHC) no Brasil. Este pacote foi criado para abordar os desafios inerentes à manipulação de grandes volumes de dados heterogêneos, característicos dos registros hospitalares de câncer, proporcionando uma ferramenta eficiente, reprodutível e acessível para pesquisadores e profissionais de saúde.
}
\details{
Veja o guia passo a passo sobre como baixar os arquivos DBF no IntegradorRHC: \code{\link{Baixar_Arquivos_DBF}}
}
\seealso{
Tutorial: \code{\link{Baixar_Arquivos_DBF}}

Reportar Bugs: \url{https://github.com/andersonlineu/RHCgen/issues}
}
\author{
Anderson Lineu

Email: \email{andersonlineu@gmail.com}

Lattes: \url{http://lattes.cnpq.br/9727490230028203}

Orcid iD: \url{https://orcid.org/0000-0002-1703-9310}

Links úteis:

Todos os arquivos DBF do RHC 1985-1999:  \url{https://1drv.ms/f/s!Al8LbsUUcnUJiPwx8DeFbAkG_YMQfw?e=JXhSgt}

Todos os arquivos DBF do RHC 2000-2022:  \url{https://1drv.ms/f/s!Al8LbsUUcnUJiJsttNv4HVyHLn7sUg?e=PMWmsL}

IntegradorRHC:  \url{https://irhc.inca.gov.br/RHCNet/}

Dicionário de Variáveis: \url{https://irhc.inca.gov.br/RHCNet/documentosTabWin.action?doc=dicionario.dados}

Escore de Completude:  \url{https://www.scielo.br/j/csp/a/HXqrdksBsMrr4R9Ydnnmqcf/?format=pdf&lang=pt}
}
