% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHCgen-package.R
\docType{package}
\name{RHCgen-package}
\alias{RHCgen-package}
\alias{RHCgen}
\title{RHCgen: Pacote para construção de Bando de Dados de registros de câncer do RHC}
\description{
RHCgen é uma biblioteca desenvolvida para facilitar a construção de bancos de dados a partir das informações sobre registros de câncer disponíveis no Registro Hospitalar de Câncer (RHC).
}
\details{
Este pacote oferece ferramentas para carregar e combinar arquivos DBF, realizar ajustes nas variáveis e, por fim, criar um banco de dados organizado para análise posterior.

Veja o Baixar_Arquivos_DBF para um guia passo a passo sobre como baixar os arquivos DBF no IntegradorRHC: \code{\link{Baixar_Arquivos_DBF}}
}
\seealso{
\code{\link{Baixar_Arquivos_DBF}}

Links úteis:

Dicionário de Variáveis: \url{ https://irhc.inca.gov.br/RHCNet/documentosTabWin.action?doc=dicionario.dados}

Escore de Completude:  \url{ https://www.scielo.br/j/csp/a/HXqrdksBsMrr4R9Ydnnmqcf/?format=pdf&lang=pt}
}
\author{
Anderson Lineu

Email: \email{andersonlineu@gmail.com}

Lattes: \url{ http://lattes.cnpq.br/9727490230028203}

Orcid iD: \url{ https://orcid.org/0000-0002-1703-9310}
}
\keyword{package}
