% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verificar_datas_inconsistentes.R
\name{verificar_datas_inconsistentes}
\alias{verificar_datas_inconsistentes}
\title{Verifica Inconsistências Entre Datas em um Dataframe}
\usage{
verificar_datas_inconsistentes(data)
}
\arguments{
\item{data}{Um dataframe contendo as colunas "Data_Diagnostico", "Data_Primeira_Consulta", "Data_Triagem" e "Data_Inicio_Primeiro_Tratamento".}
}
\value{
Retorna um dataframe com o quantitativo e a proporção de inconsistências por verificação.
}
\description{
Esta função verifica se há inconsistências na ordem cronológica das datas fornecidas em um dataframe. As verificações incluem:
\itemize{
\item "Data_Diagnostico" deve ser anterior ou igual a "Data_Primeira_Consulta"
\item "Data_Diagnostico" deve ser anterior ou igual a "Data_Triagem"
\item "Data_Inicio_Primeiro_Tratamento" deve ser anterior ou igual a "Data_Primeira_Consulta"
\item "Data_Inicio_Primeiro_Tratamento" deve ser anterior ou igual a "Data_Triagem"
\item "Data_Inicio_Primeiro_Tratamento" deve ser anterior ou igual a "Data_Diagnostico"
A função considera uma diferença mínima de 30 dias para ser considerada inconsistente.
}
}
\examples{
# Escreva o nome do dataframe e execute a função. Se seu dataframe for "dados_RHC_combinados", use a função como:

resultados_datas_inconsistentes <- verificar_datas_inconsistentes(dados_RHC_combinados)
}
